merge: Introduce "auto rebase" configuration option
Up to now, `cogite pr merge` would gracefully update the local
destination branch before rebasing and merging the working branch.
This is not always appropriate.

A new `merge_auto_rebase` option is introduced that can have 3 values:
"ask", "always", or "never".

If the local destination branch is up-to-date (with respect to the
upstream destination branch), nothing changes: Cogite merges the
working branch on the destination branch, and pushes it upstream.

If the local destination branch is NOT up-to-date, the behaviour
depends on the value of the `merge_auto_rebase` option:

- "ask" (the default value): as the name suggest, Cogite will ask for
  confirmation before updating the local destination branch, rebasing
  the working branch and proceeding with the merge.

- "always": Cogite will do as above without asking for confirmation.

- "never": Cogite will stop right there. The local destination branch
  will not be updated and the working branch will not be rebased (and,
  obviously, it will not be merged).